requirements.txt


python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0


openpyxl
python-decouple==3.5
django-admin-relation-links
zstd
async-timeout
psycopg2
djangorestframework
django-jsonform
pyzstd
python-decouple